Add html and js
Build simple html. The button is not part of
the form because when the form is submitted the
page reloads and all the changes made by js are
deleted. So I made an event listener for the
button which includes a check with modulor if the
entered number can be divided by 3 or 5.